# exmaples/example/doc/RuleEngine_swagger.json
{
    "consumes": [
        "application/json",
        "text/xml",
        "text/plain",
        "text/html",
        "application/x-www-form-urlencoded"
    ],
    "swagger": "2.0",
    "info": {
        "description": "rule engine api document，restful api",
        "title": "rule engine swagger api specification",
        "contact": {
            "name": "zhangyongxiang",
            "email": "zhangyongxiang@baidu.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/soe/rule-engine/sources": {
            "get": {
                "description": "自定义Source列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "自定义Source列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/_.SourceListElement"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/workspaces/{workspaceID}/projects/{projectName}/analyticsjobs": {
            "post": {
                "description": "普通函数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "function"
                ],
                "summary": "普通函数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspaceId",
                        "name": "workspaceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "projectName",
                        "name": "projectName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "_.SourceListElement": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}